@@ -29,7 +29,7 @@
     "prettier": "^3.1.1",
     "prettier-plugin-svelte": "^3.1.2",
     "prettier-plugin-tailwindcss": "^0.5.9",
-    "svelte": "^4.2.8",
+    "svelte": "5.0.0-next.26",
     "svelte-check": "^3.6.2",
     "tailwindcss": "^3.4.0",
     "typescript": "^5.3.3",
